[c_api] Break C API monolith into multiple targets.
