+++ xls/public/c_api.h
@@ -16,7 +16,8 @@
 #define XLS_PUBLIC_C_API_H_
 
 #include <stddef.h>  // NOLINT(modernize-deprecated-headers)
-#include <stdint.h>  // NOLINT(modernize-deprecated-headers)
+
+#include "xls/public/c_api_vast.h"
 
 // C API that exposes the functionality in various public headers in a way that
 // C-based FFI facilities can easily wrap.
@@ -91,19 +92,6 @@ bool xls_value_to_string(const struct xls_value* v, char** string_out);
 
 // Returns whether `v` is equal to `w`.
 bool xls_value_eq(const struct xls_value* v, const struct xls_value* w);
-
-// Note: We define the format preference enum with a fixed width integer type
-// for clarity of the exposed ABI.
-typedef int32_t xls_format_preference;
-enum {
-  xls_format_preference_default,
-  xls_format_preference_binary,
-  xls_format_preference_signed_decimal,
-  xls_format_preference_unsigned_decimal,
-  xls_format_preference_hex,
-  xls_format_preference_plain_binary,
-  xls_format_preference_plain_hex,
-};
 
 // Returns a format preference enum value from a string specifier; i.e.
 // `xls_format_preference_from_string("hex")` returns the value of
@@ -188,120 +176,6 @@ bool xls_function_type_to_string(struct xls_function_type* xls_function_type,
 bool xls_interpret_function(struct xls_function* function, size_t argc,
                             const struct xls_value** args, char** error_out,
                             struct xls_value** result_out);
-
-// -- VAST (Verilog AST) APIs
-//
-// Note that these are expected to be *less* stable than the above APIs, as
-// they are exposing a useful implementation library present within XLS.
-//
-// Per usual, in a general sense, no promises are made around API or ABI
-// stability overall. However, seems worth noting these are effectively
-// "protected" APIs, use with particular caution around stability. See
-// `xls/protected/BUILD` for how we tend to think about "protected" APIs in the
-// project.
-
-// Opaque structs.
-struct xls_vast_verilog_file;
-struct xls_vast_verilog_module;
-struct xls_vast_node;
-struct xls_vast_expression;
-struct xls_vast_logic_ref;
-struct xls_vast_data_type;
-struct xls_vast_indexable_expression;
-struct xls_vast_slice;
-struct xls_vast_literal;
-struct xls_vast_instantiation;
-struct xls_vast_continuous_assignment;
-
-// Note: We define the enum with a fixed width integer type for clarity of the
-// exposed ABI.
-typedef int32_t xls_vast_file_type;
-enum {
-  xls_vast_file_type_verilog,
-  xls_vast_file_type_system_verilog,
-};
-
-// Note: caller owns the returned verilog file object, to be freed by
-// `xls_vast_verilog_file_free`.
-struct xls_vast_verilog_file* xls_vast_make_verilog_file(
-    xls_vast_file_type file_type);
-
-void xls_vast_verilog_file_free(struct xls_vast_verilog_file* f);
-
-struct xls_vast_verilog_module* xls_vast_verilog_file_add_module(
-    struct xls_vast_verilog_file* f, const char* name);
-
-struct xls_vast_data_type* xls_vast_verilog_file_make_scalar_type(
-    struct xls_vast_verilog_file* f);
-
-struct xls_vast_data_type* xls_vast_verilog_file_make_bit_vector_type(
-    struct xls_vast_verilog_file* f, int64_t bit_count, bool is_signed);
-
-void xls_vast_verilog_module_add_member_instantiation(
-    struct xls_vast_verilog_module* m, struct xls_vast_instantiation* member);
-void xls_vast_verilog_module_add_member_continuous_assignment(
-    struct xls_vast_verilog_module* m,
-    struct xls_vast_continuous_assignment* member);
-
-struct xls_vast_logic_ref* xls_vast_verilog_module_add_input(
-    struct xls_vast_verilog_module* m, const char* name,
-    struct xls_vast_data_type* type);
-struct xls_vast_logic_ref* xls_vast_verilog_module_add_output(
-    struct xls_vast_verilog_module* m, const char* name,
-    struct xls_vast_data_type* type);
-struct xls_vast_logic_ref* xls_vast_verilog_module_add_wire(
-    struct xls_vast_verilog_module* m, const char* name,
-    struct xls_vast_data_type* type);
-// TODO(cdleary): 2024-09-05 Add xls_vast_verilog_module_add_wire_with_expr
-
-struct xls_vast_continuous_assignment*
-xls_vast_verilog_file_make_continuous_assignment(
-    struct xls_vast_verilog_file* f, struct xls_vast_expression* lhs,
-    struct xls_vast_expression* rhs);
-
-struct xls_vast_instantiation* xls_vast_verilog_file_make_instantiation(
-    struct xls_vast_verilog_file* f, const char* module_name,
-    const char* instance_name, const char** parameter_port_names,
-    struct xls_vast_expression** parameter_expressions, size_t parameter_count,
-    const char** connection_port_names,
-    struct xls_vast_expression** connection_expressions,
-    size_t connection_count);
-
-void xls_vast_verilog_file_add_include(struct xls_vast_verilog_file* f,
-                                       const char* path);
-
-struct xls_vast_slice* xls_vast_verilog_file_make_slice_i64(
-    struct xls_vast_verilog_file* f,
-    struct xls_vast_indexable_expression* subject, int64_t hi, int64_t lo);
-
-struct xls_vast_literal* xls_vast_verilog_file_make_plain_literal(
-    struct xls_vast_verilog_file* f, int32_t value);
-
-// Creates a VAST literal with an arbitrary bit count.
-//
-// Returns an error if the given format preference is invalid.
-bool xls_vast_verilog_file_make_literal(struct xls_vast_verilog_file* f,
-                                        struct xls_bits* bits,
-                                        xls_format_preference format_preference,
-                                        bool emit_bit_count, char** error_out,
-                                        struct xls_vast_literal** literal_out);
-
-// Casts to turn the given node to an expression, where possible.
-struct xls_vast_expression* xls_vast_literal_as_expression(
-    struct xls_vast_literal* v);
-struct xls_vast_expression* xls_vast_logic_ref_as_expression(
-    struct xls_vast_logic_ref* v);
-struct xls_vast_expression* xls_vast_slice_as_expression(
-    struct xls_vast_slice* v);
-
-struct xls_vast_indexable_expression*
-xls_vast_logic_ref_as_indexable_expression(
-    struct xls_vast_logic_ref* logic_ref);
-
-// Emits/formats the contents of the given verilog file to a string.
-//
-// Note: caller owns the returned string, to be freed by `xls_c_str_free`.
-char* xls_vast_verilog_file_emit(const struct xls_vast_verilog_file* f);
 
 }  // extern "C"
